move keyed tests to their own file
